refactor(autodiff): centralize GradTensor ops and clarify API
