+++ crates/bolt-autodiff/src/graph.rs
@@ -63,11 +63,11 @@ where
         }
     }
 
-    pub fn param<'g>(&'g self, tensor: &Tensor<B, D>) -> GradTensor<'g, B, D> {
+    pub fn variable<'g>(&'g self, tensor: &Tensor<B, D>) -> GradTensor<'g, B, D> {
         self.create_node(tensor.clone(), true, true, ArrayVec::new(), None)
     }
 
-    pub fn input<'g>(&'g self, tensor: &Tensor<B, D>) -> GradTensor<'g, B, D> {
+    pub fn constant<'g>(&'g self, tensor: &Tensor<B, D>) -> GradTensor<'g, B, D> {
         self.create_node(tensor.clone(), false, true, ArrayVec::new(), None)
     }
