+++ crates/bolt-autodiff/src/lib.rs
@@ -3,16 +3,16 @@
 mod backward;
 pub mod error;
 mod float;
+mod grad_tensor;
 mod gradients;
 mod graph;
 mod handle;
 pub mod ops;
-mod tensor;
 
 pub use backward::{BackwardContext, BackwardOp};
 pub use error::{Error, Result};
 pub use float::Float;
+pub use grad_tensor::{GradTensor, TensorLike};
 pub use gradients::Gradients;
 pub use graph::{Graph, NoGradGuard};
 pub use handle::Handle;
-pub use tensor::{Attach, AttachBuilder, GradTensor, TensorLike};
