+++ crates/bolt-autodiff/src/ops/binary.rs
@@ -5,7 +5,7 @@ use bolt_core::{
 use tinyvec::ArrayVec;
 
 use crate::{
-    Float, GradTensor, TensorLike,
+    Float,
     backward::{BackwardContext, BackwardOp, MAX_INPUTS},
     error::Result,
     ops::reduce_grad_to_shape,
@@ -174,177 +174,5 @@ where
 
     fn name(&self) -> &'static str {
         "MatmulBackward"
-    }
-}
-
-impl<'g, B, D> GradTensor<'g, B, D>
-where
-    B: Backend<D> + AddOp<D> + CopyOp<D> + SumOp<D>,
-    D: Float,
-{
-    pub fn add<T>(&self, other: T) -> Result<GradTensor<'g, B, D>>
-    where
-        T: TensorLike<'g, B, D>,
-    {
-        let self_tensor = self.tensor()?;
-        let other_node = other.into_node(self.graph());
-        let other_tensor = other_node.tensor()?;
-
-        let result = self_tensor.add(&other_tensor)?;
-
-        let self_requires_grad = self.requires_grad()?;
-        let other_requires_grad = other_node.requires_grad()?;
-        let requires_grad =
-            (self_requires_grad || other_requires_grad) && self.graph().is_grad_enabled();
-
-        if !requires_grad {
-            return Ok(self.graph().input(&result));
-        }
-
-        let saved_idx = self.graph().save_tensors_for_backward(vec![]);
-        let backward_op =
-            AddBackward::new(self_tensor.shape().to_vec(), other_tensor.shape().to_vec());
-
-        let mut inputs = ArrayVec::new();
-        inputs.push(self.handle());
-        inputs.push(other_node.handle());
-
-        Ok(self.graph().create_node(
-            result,
-            true,
-            false,
-            inputs,
-            Some((Box::new(backward_op), saved_idx)),
-        ))
-    }
-}
-
-impl<'g, B, D> GradTensor<'g, B, D>
-where
-    B: Backend<D> + AddOp<D> + SubOp<D> + FillOp<D> + CopyOp<D> + SumOp<D>,
-    D: Float,
-{
-    pub fn sub<T>(&self, other: T) -> Result<GradTensor<'g, B, D>>
-    where
-        T: TensorLike<'g, B, D>,
-    {
-        let self_tensor = self.tensor()?;
-        let other_node = other.into_node(self.graph());
-        let other_tensor = other_node.tensor()?;
-
-        let result = self_tensor.sub(&other_tensor)?;
-
-        let self_requires_grad = self.requires_grad()?;
-        let other_requires_grad = other_node.requires_grad()?;
-        let requires_grad =
-            (self_requires_grad || other_requires_grad) && self.graph().is_grad_enabled();
-
-        if !requires_grad {
-            return Ok(self.graph().input(&result));
-        }
-
-        let saved_idx = self.graph().save_tensors_for_backward(vec![]);
-        let backward_op =
-            SubBackward::new(self_tensor.shape().to_vec(), other_tensor.shape().to_vec());
-
-        let mut inputs = ArrayVec::new();
-        inputs.push(self.handle());
-        inputs.push(other_node.handle());
-
-        Ok(self.graph().create_node(
-            result,
-            true,
-            false,
-            inputs,
-            Some((Box::new(backward_op), saved_idx)),
-        ))
-    }
-}
-
-impl<'g, B, D> GradTensor<'g, B, D>
-where
-    B: Backend<D> + AddOp<D> + MulOp<D> + CopyOp<D> + SumOp<D>,
-    D: Float,
-{
-    pub fn mul<T>(&self, other: T) -> Result<GradTensor<'g, B, D>>
-    where
-        T: TensorLike<'g, B, D>,
-    {
-        let self_tensor = self.tensor()?;
-        let other_node = other.into_node(self.graph());
-        let other_tensor = other_node.tensor()?;
-
-        let result = self_tensor.mul(&other_tensor)?;
-
-        let self_requires_grad = self.requires_grad()?;
-        let other_requires_grad = other_node.requires_grad()?;
-        let requires_grad =
-            (self_requires_grad || other_requires_grad) && self.graph().is_grad_enabled();
-
-        if !requires_grad {
-            return Ok(self.graph().input(&result));
-        }
-
-        let saved_idx = self
-            .graph()
-            .save_tensors_for_backward(vec![self_tensor.clone(), other_tensor.clone()]);
-        let backward_op =
-            MulBackward::new(self_tensor.shape().to_vec(), other_tensor.shape().to_vec());
-
-        let mut inputs = ArrayVec::new();
-        inputs.push(self.handle());
-        inputs.push(other_node.handle());
-
-        Ok(self.graph().create_node(
-            result,
-            true,
-            false,
-            inputs,
-            Some((Box::new(backward_op), saved_idx)),
-        ))
-    }
-}
-
-impl<'g, B, D> GradTensor<'g, B, D>
-where
-    B: Backend<D> + AddOp<D> + MatmulOp<D> + CopyOp<D> + SumOp<D>,
-    D: Float,
-{
-    pub fn matmul<T>(&self, other: T) -> Result<GradTensor<'g, B, D>>
-    where
-        T: TensorLike<'g, B, D>,
-    {
-        let self_tensor = self.tensor()?;
-        let other_node = other.into_node(self.graph());
-        let other_tensor = other_node.tensor()?;
-
-        let result = self_tensor.matmul(&other_tensor)?;
-
-        let self_requires_grad = self.requires_grad()?;
-        let other_requires_grad = other_node.requires_grad()?;
-        let requires_grad =
-            (self_requires_grad || other_requires_grad) && self.graph().is_grad_enabled();
-
-        if !requires_grad {
-            return Ok(self.graph().input(&result));
-        }
-
-        let saved_idx = self
-            .graph()
-            .save_tensors_for_backward(vec![self_tensor.clone(), other_tensor.clone()]);
-        let backward_op =
-            MatmulBackward::new(self_tensor.shape().to_vec(), other_tensor.shape().to_vec());
-
-        let mut inputs = ArrayVec::new();
-        inputs.push(self.handle());
-        inputs.push(other_node.handle());
-
-        Ok(self.graph().create_node(
-            result,
-            true,
-            false,
-            inputs,
-            Some((Box::new(backward_op), saved_idx)),
-        ))
     }
 }
